test/buildcache: run recover check on native
iwasm does not provide the setjmp/longjmp imports used by LLGo defer. Keep the WASM fixture free of that defer path, and execute native artifacts after every cache scenario so dependency archive hits still validate cross-package recover.
