+++ test/buildcache/dep1/recover.go
@@ -1,0 +1,8 @@
+//go:build llgo && !wasip1
+
+package dep1
+
+// Recover stores the panic value seen by a directly deferred dependency call.
+func Recover(dst *any) {
+	*dst = recover()
+}
