+++ test/buildcache/main.go
@@ -12,14 +12,7 @@ import (
 )
 
 func main() {
-	var recovered any
-	func() {
-		defer dep1.Recover(&recovered)
-		panic("cached recover")
-	}()
-	if recovered != "cached recover" {
-		panic("dependency recover failed after cache lookup")
-	}
+	verifyRecoverCache()
 
 	c.Printf(c.Str("dep1.Add(1,2) = %d\n"), dep1.Add(1, 2))
 	c.Printf(c.Str("dep2.Double(1,2) = %d\n"), dep2.Double(1, 2))
